Iniciado el adaptado a la nueva base de datos

+++ app/Http/Controllers/UsuariosController.php
@@ -13,115 +13,102 @@ class UsuariosController extends Controller
     public function MenuUsuarios(Request $request){
         $usuarios = Usuario::all();
         foreach($usuarios as $usuario){
-            $usuario->isAdmin = $usuario->admin()->exists();
-            $usuario->isAlmacenero = $usuario->almacenero()->exists();
-            $usuario->isChofer = $usuario->chofer()->exists();
+            $usuario->isAdmin = !is_null(Admin::find($usuario->id));
+            $usuario->isAlmacenero = !is_null(Almacenero::find($usuario->id));
+            $usuario->isChofer = !is_null(Chofer::find($usuario->id));
         }
         return view('gestionUsuarios', ['usuarios' => $usuarios]);
     }
     public function CrearUsuario(Request $request){
         $usuario = new Usuario();
-        $usuario->nombre_usuario = $request->post('nombreUsuario');
-        $usuario->contrasena = password_hash($request->post('contrasena'), PASSWORD_DEFAULT);
+        $usuario->username = $request->post('username');
+        $usuario->password = password_hash($request->post('password'), PASSWORD_DEFAULT);
         $usuario->nombre = $request->post('nombre');
         $usuario->apellido = $request->post('apellido');
-        $usuario->email = $request->post('email');
-        $usuario->idioma_favorito = $request->post('idiomaFavorito');
-        $usuario->codigo_postal = $request->post('codigoPostal');
-        $usuario->calle = $request->post('calle');
-        $usuario->nro_puerta = $request->post('nroPuerta');
-        $usuario->observaciones_direccion = $request->post('observacionesDireccion');
-        $usuario->latitud = $request->post('latitud');
-        $usuario->longitud = $request->post('longitud');
         $usuario->save();
         if($request->post('admin')){
             $admin = new Admin();
-            $admin->nombre_usuario = $usuario->nombre_usuario;
+            $admin->id = $usuario->id;
             $admin->save();
         }
         if($request->post('chofer')){
             $chofer = new Chofer();
-            $chofer->nombre_usuario = $usuario->nombre_usuario;
+            $chofer->id = $usuario->id;
             $chofer->save();
         }
         if($request->post('almacenero')){
             $almacenero = new Almacenero();
-            $almacenero->nombre_usuario = $usuario->nombre_usuario;
+            $almacenero->id = $usuario->id;
             $almacenero->save();
         }
-        return redirect('/');
+        return json_encode($usuario);
     }
+
     public function BorrarUsuario(Request $request){
-        $usuario = Usuario::find($request->nombre_usuario);
+        $usuario = Usuario::findOrFail($request->id);
         
-        if($usuario->almacenero()->exists()){
-            $almacenero = Almacenero::find($request->nombre_usuario);
+        if(Almacenero::find($request->id)){
+            $almacenero = Almacenero::find($request->id);
             $almacenero->delete();
         }
         
-        if($usuario->admin()->exists()){
-        $admin = Admin::find($request->nombre_usuario);
-        $admin->delete();
+        if(Admin::find($request->id)){
+            $admin = Admin::find($request->id);
+            $admin->delete();
         }
 
-        if($usuario->chofer()->exists()){
-        $chofer = Chofer::find($request->nombre_usuario);
-        $chofer->delete();
+        if(Chofer::find($request->id)){
+            $chofer = Chofer::find($request->id);
+            $chofer->delete();
         }
         
         $usuario->delete();
 
         return redirect('/');
     }
+
     public function CrearFormulario(Request $request){
-        $usuario = Usuario::find($request->nombre_usuario);
+        $usuario = Usuario::find($request->id);
         $usuario->isAdmin = $usuario->admin()->exists();
         $usuario->isAlmacenero = $usuario->almacenero()->exists();
         $usuario->isChofer = $usuario->chofer()->exists();
         return view('formularioUsuarios', ["usuario" => $usuario]);
     }
+
     public function UpdateUsuarios(Request $request){
-        $usuario = Usuario::find($request->nombre_usuario);
-        $usuario->nombre_usuario = $request->post('nombreUsuario');
-        $usuario->contrasena = password_hash($request->post('contrasena'), PASSWORD_DEFAULT);
+        $usuario = Usuario::find($request->id);
+        $usuario->username = $request->post('username');
+        $usuario->password = password_hash($request->post('password'), PASSWORD_DEFAULT);
         $usuario->nombre = $request->post('nombre');
         $usuario->apellido = $request->post('apellido');
-        $usuario->email = $request->post('email');
-        $usuario->idioma_favorito = $request->post('idiomaFavorito');
-        $usuario->codigo_postal = $request->post('codigoPostal');
-        $usuario->calle = $request->post('calle');
-        $usuario->nro_puerta = $request->post('nroPuerta');
-        $usuario->observaciones_direccion = $request->post('observacionesDireccion');
-        $usuario->latitud = $request->post('latitud');
-        $usuario->longitud = $request->post('longitud');
         $usuario->save();
         if($usuario->almacenero()->exists()){
-            $almacenero = Almacenero::find($request->nombre_usuario);
+            $almacenero = Almacenero::find($request->id);
             $almacenero->delete();
         }
         
         if($usuario->admin()->exists()){
-        $admin = Admin::find($request->nombre_usuario);
+        $admin = Admin::find($request->id);
         $admin->delete();
         }
 
         if($usuario->chofer()->exists()){
-        $chofer = Chofer::find($request->nombre_usuario);
+        $chofer = Chofer::find($request->id);
         $chofer->delete();
         }
         if($request->post('admin')){
             $admin = new Admin();
-            $admin->nombre_usuario = $usuario->nombre_usuario;
+            $admin->id = $usuario->id;
             $admin->save();
         }
         if($request->post('chofer')){
             $chofer = new Chofer();
-            $chofer->nombre_usuario = $usuario->nombre_usuario;
+            $chofer->id = $usuario->id;
             $chofer->save();
         }
         if($request->post('almacenero')){
             $almacenero = new Almacenero();
-            $almacenero->nombre_usuario = $usuario->nombre_usuario;
+            $almacenero->id = $usuario->id;
             $almacenero->save();
         }
         return redirect('/');
